size_t to int

+++ get_next_line.h
@@ -29,9 +29,9 @@
 int		get_next_line(int fd, char **line);
 
 char	*ft_strchr(char *s, int c);
-size_t	ft_strlen(char *c);
-char	*ft_strcpy(char	*dst, char *stc, size_t n);
-char	*ft_strndup(char *s1, size_t size);
+int	    ft_strlen(char *c);
+char	*ft_strcpy(char	*dst, char *stc, int n);
+char	*ft_strndup(char *s1, int size);
 char	*ft_strjoin(char *s1, char *s2);
 
 #endif
